+++ Main Apk/Main Apk-BACKEND/SecurityNewApk/src/main/java/com/example/newsecurity/Controller/SystemAdministratorController.java
@@ -1,17 +1,17 @@
 package com.example.newsecurity.Controller;
 
-import com.example.newsecurity.DTO.ManagerUpdateDTO;
 import com.example.newsecurity.DTO.SystemAdministratorUpdateDTO;
-import com.example.newsecurity.Model.Manager;
 import com.example.newsecurity.Model.SystemAdministrator;
+import com.example.newsecurity.Model.User;
 import com.example.newsecurity.Repository.ISystemAdministratorRepository;
 import com.example.newsecurity.Service.ISystemAdministratorService;
+import com.example.newsecurity.Service.IUserService;
+import com.example.newsecurity.Util.TokenUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
-import java.util.NoSuchElementException;
 
 @RestController
 @RequestMapping(value = "/system-administrators",produces = MediaType.APPLICATION_JSON_VALUE)
@@ -21,26 +21,53 @@ public class SystemAdministratorController {
     private ISystemAdministratorService systemAdministratorService;
     @Autowired
     private ISystemAdministratorRepository systemAdministratorRepository;
+    @Autowired
+    private TokenUtils tokenUtils;
+    @Autowired
+    private IUserService userService;
 
     @GetMapping
-    public List<SystemAdministrator> getAllAdministrators(){
+    public List<SystemAdministrator> getAllAdministrators(@RequestHeader("Authorization") String authorizationHeader){
+        String jwtToken = authorizationHeader.replace("Bearer ", "");
+        String username = tokenUtils.getUsernameFromToken(jwtToken);
+        User user = userService.findByUsername(username);
+        if (!user.hasPermission("GET_ADMIN_ALL")){
+            return null;
+        }
         return systemAdministratorService.getAllAdministrators();
     }
 
     @GetMapping("/{id}")
-    public SystemAdministrator getAdministratorById(@PathVariable Long id){
+    public SystemAdministrator getAdministratorById(@RequestHeader("Authorization") String authorizationHeader, @PathVariable Long id){
+        String jwtToken = authorizationHeader.replace("Bearer ", "");
+        String username = tokenUtils.getUsernameFromToken(jwtToken);
+        User user = userService.findByUsername(username);
+        if (!user.hasPermission("GET_ADMIN_ONE")){
+            return null;
+        }
         return systemAdministratorService.getAdministratorById(id);
     }
 
     @DeleteMapping("/{id}")
-    public String deleteAdministratorById(@PathVariable Long id){
+    public String deleteAdministratorById(@RequestHeader("Authorization") String authorizationHeader, @PathVariable Long id){
+        String jwtToken = authorizationHeader.replace("Bearer ", "");
+        String username = tokenUtils.getUsernameFromToken(jwtToken);
+        User user = userService.findByUsername(username);
+        if (!user.hasPermission("DELETE_ADMIN")){
+            return null;
+        }
         systemAdministratorService.deleteAdministratorById(id);
         return "System administrator deleted successfully!";
     }
 
     @PutMapping
-    public SystemAdministrator updateAdministrator(@RequestBody SystemAdministratorUpdateDTO adminDTO){
+    public SystemAdministrator updateAdministrator(@RequestHeader("Authorization") String authorizationHeader, @RequestBody SystemAdministratorUpdateDTO adminDTO){
+        String jwtToken = authorizationHeader.replace("Bearer ", "");
+        String username = tokenUtils.getUsernameFromToken(jwtToken);
+        User user = userService.findByUsername(username);
+        if (!user.hasPermission("UPDATE_ADMIN")){
+            return null;
+        }
         return systemAdministratorService.updateAdministrator(adminDTO);
-
     }
 }
